+++ app/routes/import-message.js
@@ -9,7 +9,8 @@ const getConvoMiddleware = require('../../lib/middleware/conversation-get');
 const createConvoMiddleware = require('../../lib/middleware/conversation-create');
 const broadcastMiddleware = require('../../lib/middleware/import-message/broadcast');
 const updateConvoMiddleware = require('../../lib/middleware/import-message/conversation-update');
-const outboundMessageMiddleware = require('../../lib/middleware/import-message/message-outbound');
+const loadOutboundImportMessageMiddleware = require('../../lib/middleware/import-message/message-outbound-load');
+const createOutboundImportMessageMiddleware = require('../../lib/middleware/import-message/message-outbound-create');
 
 router.use(paramsMiddleware());
 router.use(broadcastMiddleware());
@@ -20,7 +21,8 @@ router.use(createConvoMiddleware());
 
 router.use(updateConvoMiddleware());
 
-// Create outbound message
-router.use(outboundMessageMiddleware());
+// Load/create outbound message
+router.use(loadOutboundImportMessageMiddleware());
+router.use(createOutboundImportMessageMiddleware());
 
 module.exports = router;
